+++ opentelemetry-otlp/examples/basic-otlp/src/main.rs
@@ -59,9 +59,12 @@ fn init_logs() -> SdkLoggerProvider {
         .build()
 }
 
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
-    let logger_provider = init_logs();
+// #[tokio::main]
+fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+    let rt = tokio::runtime::Runtime::new()?;
+     let logger_provider = rt.block_on(async {
+        init_logs()
+     });
 
     // Create a new OpenTelemetryTracingBridge using the above LoggerProvider.
     let otel_layer = OpenTelemetryTracingBridge::new(&logger_provider);
@@ -98,66 +101,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
         .with(otel_layer)
         .with(fmt_layer)
         .init();
-
-    // At this point Logs (OTel Logs and Fmt Logs) are initialized, which will
-    // allow internal-logs from Tracing/Metrics initializer to be captured.
-
-    let tracer_provider = init_traces();
-    // Set the global tracer provider using a clone of the tracer_provider.
-    // Setting global tracer provider is required if other parts of the application
-    // uses global::tracer() or global::tracer_with_version() to get a tracer.
-    // Cloning simply creates a new reference to the same tracer provider. It is
-    // important to hold on to the tracer_provider here, so as to invoke
-    // shutdown on it when application ends.
-    global::set_tracer_provider(tracer_provider.clone());
-
-    let meter_provider = init_metrics();
-    // Set the global meter provider using a clone of the meter_provider.
-    // Setting global meter provider is required if other parts of the application
-    // uses global::meter() or global::meter_with_version() to get a meter.
-    // Cloning simply creates a new reference to the same meter provider. It is
-    // important to hold on to the meter_provider here, so as to invoke
-    // shutdown on it when application ends.
-    global::set_meter_provider(meter_provider.clone());
-
-    let common_scope_attributes = vec![KeyValue::new("scope-key", "scope-value")];
-    let scope = InstrumentationScope::builder("basic")
-        .with_version("1.0")
-        .with_attributes(common_scope_attributes)
-        .build();
-
-    let tracer = global::tracer_with_scope(scope.clone());
-    let meter = global::meter_with_scope(scope);
-
-    let counter = meter
-        .u64_counter("test_counter")
-        .with_description("a simple counter for demo purposes.")
-        .with_unit("my_unit")
-        .build();
-    for _ in 0..10 {
-        counter.add(1, &[KeyValue::new("test_key", "test_value")]);
-    }
-
-    tracer.in_span("Main operation", |cx| {
-        let span = cx.span();
-        span.add_event(
-            "Nice operation!".to_string(),
-            vec![KeyValue::new("bogons", 100)],
-        );
-        span.set_attribute(KeyValue::new("another.key", "yes"));
-
-        info!(name: "my-event-inside-span", target: "my-target", "hello from {}. My price is {}. I am also inside a Span!", "banana", 2.99);
-
-        tracer.in_span("Sub operation...", |cx| {
-            let span = cx.span();
-            span.set_attribute(KeyValue::new("another.key", "yes"));
-            span.add_event("Sub span event", vec![]);
-        });
-    });
-
     info!(name: "my-event", target: "my-target", "hello from {}. My price is {}", "apple", 1.99);
-    tracer_provider.shutdown()?;
-    meter_provider.shutdown()?;
     logger_provider.shutdown()?;
 
     Ok(())